backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
sentence-transformers
pinecone-client
python-dotenv
google-generativeai
transformers